[整理] コードを分離し整理
