+++ .vscode/settings.json
@@ -12,6 +12,7 @@
   ],
   "cSpell.words": [
     "Akar",
+    "apac",
     "asar",
     "AUTOINCREMENT",
     "Elems",
