+++ src/errors/llm-errors.ts
@@ -11,3 +11,10 @@ export class ToolInvocationError extends Error {
     this.name = 'ToolInvocationError';
   }
 }
+
+export class InvalidParameterError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = 'InvalidParameterError';
+  }
+}
